+++ src/Api/Api.php
@@ -19,6 +19,7 @@
 namespace CloudCreativity\LaravelJsonApi\Api;
 
 use CloudCreativity\JsonApi\Contracts\Encoder\SerializerInterface;
+use CloudCreativity\JsonApi\Contracts\Http\Client\ClientInterface;
 use CloudCreativity\JsonApi\Contracts\Http\Responses\ErrorResponseInterface;
 use CloudCreativity\JsonApi\Contracts\Repositories\ErrorRepositoryInterface;
 use CloudCreativity\JsonApi\Contracts\Store\StoreInterface;
@@ -334,6 +335,15 @@ class Api
             $extensions ?: $this->getSupportedExtensions(),
             (string) $this->getUrl()
         );
+    }
+
+    /**
+     * @param $httpClient
+     * @return ClientInterface
+     */
+    public function createClient($httpClient)
+    {
+        return $this->factory->createClient($httpClient, $this->getSchemas(), $this->createEncoder());
     }
 
     /**