+++ src/Routing/ApiGroup.php
@@ -100,6 +100,7 @@ class ApiGroup
     {
         return [
             'default-authorizer' => $this->options->get('authorizer'),
+            'prefix' => $this->api->getUrl()->getNamespace(),
         ];
     }